requirements.txt
# Linting
flake8

# Jukebox
# sudo apt install libasound2-dev required on some machines
components
evdev
eyed3
mock
nanotime
numpy
paho_mqtt
py532lib # RFID Reader
pyalsaaudio
pyserial
pytest
python_mpd2
# pyzmq
# RPi.GPIO
ruamel.yaml
# smbus
tornado
requests
sphinx
sphinx_rtd_theme
